More work. Getting closer!

+++ src/integrals/active_space_integrals.h
@@ -205,7 +205,7 @@ class ActiveSpaceIntegrals {
     /// A vector of indices for the active molecular orbitals
     std::vector<size_t> active_mo_;
     /// A vector of the symmetry ofthe active molecular orbitals
-    std::vector<size_t> active_mo_symmetry_;
+    std::vector<int> active_mo_symmetry_;
     /// A Vector of indices for the restricted_docc molecular orbitals
     std::vector<size_t> restricted_docc_mo_;
